MLIBZ-2627: Fix push module to reuse main module components
