+++ packages/kinvey-nativescript-sdk/src/bundle.ts
@@ -31,6 +31,11 @@ export {
   Query,
   User,
 
+  AuthType,
+  KinveyRequest,
+  RequestMethod,
+  DataAccess,
+
   ActiveUserError,
   APIVersionNotAvailableError,
   APIVersionNotImplementedError,
