+++ packages/kinvey-nativescript-sdk/src/push/package.json
