+++ packages/kinvey-nativescript-sdk/src/push/push.android.ts
@@ -1,8 +1,7 @@
 import { Promise } from 'es6-promise';
-import { isDefined } from '../../core/utils';
-import { KinveyError } from '../../core/errors';
+import { KinveyError } from './kinvey-nativescript-sdk';
 import { PushCommon } from './common';
-import { PushConfig, IOSPushConfig } from './';
+import { PushConfig, AndroidPushConfig } from './';
 // tslint:disable-next-line:variable-name
 let PushPlugin;
 
@@ -12,46 +11,35 @@ try {
   // Just catch the error
 }
 
-class IOSPush extends PushCommon {
+class AndroidPush extends PushCommon {
   protected _registerWithPushPlugin(options = <PushConfig>{}): Promise<string> {
-    const config = options.ios || <IOSPushConfig>{};
+    const config = options.android || <AndroidPushConfig>{};
 
     return new Promise((resolve, reject) => {
-      if (isDefined(PushPlugin) === false) {
+      if (!PushPlugin) {
         return reject(new KinveyError('NativeScript Push Plugin is not installed.',
           'Please refer to http://devcenter.kinvey.com/nativescript/guides/push#ProjectSetUp for help with'
           + ' setting up your project.'));
       }
 
-      const usersNotificationCallback = config.notificationCallbackIOS;
-      config.notificationCallbackIOS = (message: any) => {
+      const usersNotificationCallback = config.notificationCallbackAndroid;
+      config.notificationCallbackAndroid = (jsonDataString: string, fcmNotification) => {
         if (typeof usersNotificationCallback === 'function') {
-          usersNotificationCallback(message);
+          usersNotificationCallback(jsonDataString, fcmNotification);
         }
 
-        (this as any).emit('notification', message);
+        (this as any).emit('notification', JSON.parse(jsonDataString));
       };
 
-      PushPlugin.register(config, (token) => {
-        if (isDefined(config.interactiveSettings)) {
-          PushPlugin.registerUserNotificationSettings(() => {
-            resolve(token);
-          }, (error) => {
-            // do something with error
-            resolve(token);
-          });
-        } else {
-          resolve(token);
-        }
-      }, reject);
+      PushPlugin.register(config, resolve, reject);
     });
   }
 
   protected _unregisterWithPushPlugin(options = <PushConfig>{}): Promise<null> {
-    const config = options.ios || <IOSPushConfig>{};
+    const config = options.android || <AndroidPushConfig>{};
 
     return new Promise((resolve, reject) => {
-      if (isDefined(PushPlugin) === false) {
+      if (!PushPlugin) {
         return reject(new KinveyError('NativeScript Push Plugin is not installed.',
           'Please refer to http://devcenter.kinvey.com/nativescript/guides/push#ProjectSetUp for help with'
           + ' setting up your project.'));
@@ -63,5 +51,5 @@ class IOSPush extends PushCommon {
 }
 
 // tslint:disable-next-line:variable-name
-const Push = new IOSPush();
+const Push = new AndroidPush();
 export { Push };
