+++ packages/kinvey-nativescript-sdk/src/push/push.ios.ts
@@ -1,8 +1,7 @@
 import { Promise } from 'es6-promise';
-import { isDefined } from '../../core/utils';
-import { KinveyError } from '../../core/errors';
+import { KinveyError } from './kinvey-nativescript-sdk';
 import { PushCommon } from './common';
-import { PushConfig, AndroidPushConfig } from './';
+import { PushConfig, IOSPushConfig } from './';
 // tslint:disable-next-line:variable-name
 let PushPlugin;
 
@@ -12,35 +11,46 @@ try {
   // Just catch the error
 }
 
-class AndroidPush extends PushCommon {
+class IOSPush extends PushCommon {
   protected _registerWithPushPlugin(options = <PushConfig>{}): Promise<string> {
-    const config = options.android || <AndroidPushConfig>{};
+    const config = options.ios || <IOSPushConfig>{};
 
     return new Promise((resolve, reject) => {
-      if (isDefined(PushPlugin) === false) {
+      if (!PushPlugin) {
         return reject(new KinveyError('NativeScript Push Plugin is not installed.',
           'Please refer to http://devcenter.kinvey.com/nativescript/guides/push#ProjectSetUp for help with'
           + ' setting up your project.'));
       }
 
-      const usersNotificationCallback = config.notificationCallbackAndroid;
-      config.notificationCallbackAndroid = (jsonDataString: string, fcmNotification) => {
+      const usersNotificationCallback = config.notificationCallbackIOS;
+      config.notificationCallbackIOS = (message: any) => {
         if (typeof usersNotificationCallback === 'function') {
-          usersNotificationCallback(jsonDataString, fcmNotification);
+          usersNotificationCallback(message);
         }
 
-        (this as any).emit('notification', JSON.parse(jsonDataString));
+        (this as any).emit('notification', message);
       };
 
-      PushPlugin.register(config, resolve, reject);
+      PushPlugin.register(config, (token) => {
+        if (!config.interactiveSettings) {
+          PushPlugin.registerUserNotificationSettings(() => {
+            resolve(token);
+          }, (error) => {
+            // do something with error
+            resolve(token);
+          });
+        } else {
+          resolve(token);
+        }
+      }, reject);
     });
   }
 
   protected _unregisterWithPushPlugin(options = <PushConfig>{}): Promise<null> {
-    const config = options.android || <AndroidPushConfig>{};
+    const config = options.ios || <IOSPushConfig>{};
 
     return new Promise((resolve, reject) => {
-      if (isDefined(PushPlugin) === false) {
+      if (!PushPlugin) {
         return reject(new KinveyError('NativeScript Push Plugin is not installed.',
           'Please refer to http://devcenter.kinvey.com/nativescript/guides/push#ProjectSetUp for help with'
           + ' setting up your project.'));
@@ -52,5 +62,5 @@ class AndroidPush extends PushCommon {
 }
 
 // tslint:disable-next-line:variable-name
-const Push = new AndroidPush();
+const Push = new IOSPush();
 export { Push };
