+++ packages/kinvey-nativescript-sdk/webpack.config.android.js
@@ -30,6 +30,7 @@ module.exports = (env = {}) => {
       'nativescript-sqlite': 'nativescript-sqlite',
       'tns-core-modules/application': 'application',
       http: 'http',
+      './kinvey-nativescript-sdk': './kinvey-nativescript-sdk',
       'tns-core-modules/http': 'http',
       'tns-core-modules/file-system': 'file-system',
       'tns-core-modules/ui/frame': 'ui/frame',
@@ -65,7 +66,7 @@ module.exports = (env = {}) => {
   bundleName = `${bundleName}.${platform}`;
   pushBundleName = `${pushBundleName}.${platform}`;
   config.entry[bundleName] = './src/index.ts';
-  config.entry[pushBundleName] = './src/push.ts';
+  config.entry[pushBundleName] = './src/push';
   return config;
 };
